feat(api): rename: Vilage to Srt
