+++ src/main/java/com/example/demo/weather/controller/WeatherController.java
@@ -5,21 +5,17 @@ import com.example.demo.weather.dto.PointDto;
 import com.example.demo.weather.dto.WeatherForecast;
 
 import com.example.demo.weather.dto.WeatherNowcast;
-import com.example.demo.weather.dto.geocoding.GeoNcpResponse;
-import com.example.demo.weather.dto.ncst.NcstApiResponse;
 import com.example.demo.weather.dto.news.NDNewsResponse;
 import com.example.demo.weather.dto.rgeocoding.RGeoResponseDto;
 import com.example.demo.weather.service.GridConversionService;
 import com.example.demo.weather.service.MidFcstService;
 import com.example.demo.weather.service.NDSearchService;
-import com.example.demo.weather.service.VilageSrtFcstService;
+import com.example.demo.weather.service.SrtFcstService;
 import com.example.demo.weather.service.NcpGeocodeService;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -28,7 +24,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("api/v1/weather")
 @RequiredArgsConstructor
 public class WeatherController {
-    private final VilageSrtFcstService vilageSrtFcstService;
+    private final SrtFcstService srtFcstService;
     private final MidFcstService midFcstService;
     private final NcpGeocodeService geocodeService;
     private final NDSearchService ndSearchService;
@@ -69,7 +65,7 @@ public class WeatherController {
             @RequestParam("ny")
             Integer ny
     ) {
-        return vilageSrtFcstService.getUltraSrtFcst(nx, ny);
+        return srtFcstService.getUltraSrtFcst(nx, ny);
     }
 
     /**
@@ -82,7 +78,7 @@ public class WeatherController {
             @RequestParam("ny")
             Integer ny
     ) {
-        return vilageSrtFcstService.getUltraSrtNcst(nx, ny);
+        return srtFcstService.getUltraSrtNcst(nx, ny);
     }
 
     /**
