+++ src/main/java/com/example/demo/weather/dto/sfcst/FcstApiResponse.java
@@ -1,4 +1,4 @@
-package com.example.demo.weather.dto.fcst;
+package com.example.demo.weather.dto.sfcst;
 
 
 import lombok.Data;
