+++ src/main/java/com/example/demo/weather/dto/sfcst/FcstItem.java
@@ -1,14 +1,15 @@
-package com.example.demo.weather.dto.ncst;
+package com.example.demo.weather.dto.sfcst;
 
-import java.util.Map;
 import lombok.Data;
 
 @Data
-public class NcstItem {
+public class FcstItem {
     private String baseDate;
     private String baseTime;
     private String category;
+    private String fcstDate;
+    private String fcstTime;
+    private String fcstValue;
     private Integer nx;
     private Integer ny;
-    private Double obsrValue;
 }
