+++ src/main/java/com/example/demo/weather/service/SrtFcstApiService.java
@@ -1,14 +1,14 @@
 package com.example.demo.weather.service;
 
-import com.example.demo.weather.dto.fcst.FcstApiResponse;
-import com.example.demo.weather.dto.ncst.NcstApiResponse;
+import com.example.demo.weather.dto.sfcst.FcstApiResponse;
+import com.example.demo.weather.dto.sncst.NcstApiResponse;
 import java.util.Map;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.service.annotation.GetExchange;
 import org.springframework.web.service.annotation.HttpExchange;
 
 @HttpExchange("/1360000/VilageFcstInfoService_2.0/")
-public interface VilageFcstApiService {
+public interface SrtFcstApiService {
 
     // 초단기 예보 조회
     @GetExchange("/getUltraSrtFcst")
